discord user ID를 BigInteger로 변경 - userID가 Integer 범위를 초과해서 BigInteger로 바꿈. - 파이썬은 int가 32비트 이상도 가능해서 DB 컬럼만 바꾸면 되더라고요...

+++ together_bot/models/fword_user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer
+from sqlalchemy import BigInteger, Column, Integer
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -12,7 +12,7 @@ class FwordUser(Base):
     __tablename__ = "fword_user"
 
     id = Column(Integer, primary_key=True)
-    discord_id = Column(Integer, unique=True)
+    discord_id = Column(BigInteger, unique=True)
 
 
 # 편의를 위해 만든 shortcut function임.
